set of changes that changes if pdf links are inserted in notes
If the calculated pdf does not exist, users are prompted for a file, or they select no pdf to add. I also changed the open notes function and how it finds the note. Unfortunately this new approach is slow for large notes files. It should be ok if notes are in individual files. I changed this because the previous approach was fragile and relied on a specific string for the custom_id, and this new approach is more robust.

This doesn't seem ideal, but I don't have a better solution right now. I don't use the notes feature much.