Rework font-lock for org-ref links
This is a major rewrite of the fontification of org-ref links. Instead
of a simple regexp, we use a search/org-element approach. This enables
all link formats to be supported, and makes it easy to put tooltips on
each link type. We also do not use highlight anymore. Now we have
tooltips on cite, label, ref, bibliographystyle and bibliography links.