Bugfix in creating notes for new bibtex entries
Function 'doi-utils-insert-bibtex-entry-from-doi' would fail to create a
note when adding a bibtex entry.
Solved by substituting a function of helm-bibtex with
'org-ref-open-bibtex-notes'

The old function would fail because the bibliography buffer was not
saved, therefore it couldn't find the requested key.
Even if fixing the error by saving the buffer before calling it, the helm-bibtex function
would write the note in a different buffer, so I decided to substitute it