fix bibliography-style issue
`orhc-formated-citations' and `or-ivy-bibtex-formatted-citation' did not
pick the bibliography style, resulting in errors when trying to insert
or copy formatted citations.

The issue seems to come from the multiple defvars for the variable
`bibliography-stile`:
1. `(defvar-local bibliography-style ...)' in org-ref-bibtex.el
2. `(defvar bibliography-style ...)' in citeproc\org-ref-citeproc.el
3. `(defvar bibliography-style ...)' in citeproc\unsrt.el,
citeproc\author-year.el...

To make it work, I removed the first defvar-local and changed the third
defvar's to setq's.