Modification of function org-ref-open-bibtex-notes
A new note in the org notes file is created according the user defined format, plus a line like:

[[cite:key]] [[pdf-path][pdf]]

Unfortunately the pdf-path is set to be absolute, so this does not work if you have several systems, for instance using org-ref on dropbox on a home and office system. I have modified the org-ref-open-bibtex notes function so the link is inserted as:

[[papers:key][pdf]]

Where papers is an org abbreviation set locally:
(setq org-link-abbrev-alist '(("papers" . "~/path/to/pdfs/%s.pdf")))

This allows org-ref to work across several systems.