Distinguish between keys and links
Here we distinguish between keys and citation links. With keys, we can run general commands associated with the key at point (aka candidate), like opening pdf, notes, tagging entries etc. I've added a keymap for those. This is probably the quickest way to do what we want. For commands that are specific to citation links, like deleting citation at point, we navigate to the links that share the same key.