New test setup
This new setup is based on src-blocks in an org-file. We can run the
tests in the org-file, or tangle them out to ert-style tests that can
run on Travis (or locally). I am getting rid of the old tests in favor
of this. I hope this encourages me (and others) to write new tests for
ort-ref.