Add gls shortcut commands
Here we use shortcuts to distinguish between glossary and acronym
links. This is because the gls command is also used for acronyms. The
problem is we would need two link types sharing the same prefix
(gls:acronym and gls:glossary), which is not possible AFAICT. One
solution is to use shortcuts:

\newacronym{lvm}{LVM}{Logical Volume Manager}
\newglossaryentry{apple}{name={apple},description={a fruit}}
ac:lvm    => \gls{lvm}
gls:apple => \gls{apple}

Note that because we export to the expanded equivalent (ac -> gls),
there's no need to specify the [shortcuts] option. See p.168 of
http://get-software.net/macros/latex/contrib/glossaries/glossaries-user.pdf