added org-ref-bibtex-assoc-pdf-with-entry
Created the function org-ref-bibtex-assoc-pdf-with-entry and added it to org-ref-bibtex-hydra. This function takes a file and will copy it to the org-ref-pdf-directory and change the name to [bibtexkey].pdf.

The use case for this is when we already have a bibtex entry for an article and we have the pdf file for that article, but the file is not named correctly for org-ref to find or it is not otherwise associated with the bibtex entry. This function makes it very easy to rapidly associate files with the bibtex entries. It is especially useful for cases where other import methods have failed to download the pdf.