update to org-ref-bibtex-assoc-pdf-with-entry
creates a defcustom for the function to be used to move the file which should be either rename-file (default) or copy-file. Adds the option to use a prefix key that will uses the opposite function to that defined by the defcustom. In other words, if you set the defcustom as read-file, adding a prefix argument will call copy-file instead, and vice versa.

Also, shortened some of the strings for the method buffer.