added function to async load helm cite sources
The first time that org-ref-helm-insert-cite-link is called,
it can take a long time to load the completion sources. This
commit creates a new function, org-ref-helm-load-completion-source-async
that will load the bibtext file into a helm source in the
background. This way, the user can continue working with their document
while helm loads their sources silently in the background.

This adds two dependecies to org-ref-helm-bibtex - async and package.
Using package withing the async call is necessary because there are
many, many functions and pacakges that go into the process of building
up the helm-source list. We could try to untangle them and only load
those that are absolutely necessary, but this would not be robust to
upstream changes.