Wrap `bibtex-completion-find-pdf' for when `helm-bibtex' isn't loaded.
If `org-ref' is started with `org-ref-completion-library' set to
anything that isn't 'helm-bibtex, then `bibtex-completion', which
contains `bibtex-completion-find-pdf', is not loaded.

This patch defines a new function `org-ref-get-pdf-filename-from-field',
which wraps `bibtex-completion-find-pdf' to load `bibtex-completion' if
it isn't already.

This handles the case of pdf files specified through
Mendeley/Jabref/Zotero without requiring that the full `helm-bibtex' be
used for everything.