remove save-buffer calls.
This was triggering some unwanted save-buffer hooks with ws-butler. I don't see
why we would need to do this unless it has to do with the org-map-entries. These
functions get called alot during font-lock to check ref labels, and every time
you change the line, they were calling save-buffer. That is not useful.