Generalized title-case conversion for entries.
This change provides a new function, org-ref-title-case, that uses the
customizable variable org-ref-title-case-types to specify what entry
types can have titles converted to title-case.

org-ref-title-case-article works exactly the same as before, but calls
org-ref-title-case with a type list limited to "article" to do the work.

Added brief notes to documentation about the changes.

Added simple test case for the changes.