Removed the check of the whole buffer in org-ref-replace-ascii.
The function org-ref-non-ascii-characters can still be used to check
the whole buffer if desired.

Modified the documentation to note that the org-ref-replace-ascii only
changes the current entry.

Added a note in the documentation about how the default behavior of
org-ref-clean-bibtex-entry can be changed by changing the initial list
of functions on the hook.