Fix issue #601.
It appears a worldcat website changed. The fix committed here is a variation on
the function proposed by @jagrg. This version keeps the same interface as
before. His solution looks like it inserts directly into the current buffer. I
don't know how common this function is used, but I wanted to keep the same signature.