Allow associated PDF files to be conveniently opened.
In issue #614 there was a request to be able to access related files from
org-ref. This commit makes it easier to access related pdf files, as long as
they are named by the convention key*.pdf, e.g. key-si.pdf, etc. The old
behavior should be preserved, e.g. if there is only one pdf, it will just get
opened, but if multiple pdfs are found matching the pattern you will now be
prompted for which pdf to open.