add new label finding code
This relies on a buffer-local list of labels that are updated incrementally by
font-lock. I think it will be much faster than the cache-parse branch, because
it does not use org-element-parse-buffer at all.

There is one currentl issue with this, which is when you open a buffer I need to
figure out how to get the whole thing fontified so that all the labels are known.