Fix minor regex mistakes
Some of these are mistakes (e.g. "\\label" instead of "\\\\label") and
others are fine but typically viewed as bad style (e.g. "*" instead of
"\\*").

Used the relint package (available on GNU ELPA) to find these. Here is
the output (all fixed):

org-ref-arxiv.el:221:51: In call to replace-regexp-in-string: Duplicated alternative branch (pos 54)
  "\"\\|\\*\\|/\\|:\\|<\\|>\\|\\?\\|\\\\\\||\\|\\+\\|,\\|\\.\\|;\\|=\\|\\[\\|]\\|:\\|!\\|@"
   ..............................................................................^
org-ref-bibtex.el:1105:14: In call to replace-regexp-in-string: Duplicated `\' inside character alternative (pos 5)
  "[\n\\|\t\\| ]+"
   ........^
org-ref-bibtex.el:1105:14: In call to replace-regexp-in-string: Duplicated `|' inside character alternative (pos 6)
  "[\n\\|\t\\| ]+"
   ..........^
org-ref-core.el:718:3: In org-ref-cite-re: Literal `-' not first or last in character alternative (pos 695)
  "\\(cite\\|nocite\\|citet\\|citet\\*\\|citep\\|citep\\*\\|citealt\\|citealt\\*\\|citealp\\|citealp\\*\\|citenum\\|citetext\\|citeauthor\\|citeauthor\\*\\|citeyear\\|citeyear\\*\\|citeyearpar\\|Citet\\|Citep\\|Citealt\\|Citealp\\|Citeauthor\\|Cite\\|parencite\\|Parencite\\|footcite\\|footcitetext\\|textcite\\|Textcite\\|smartcite\\|Smartcite\\|cite\\*\\|parencite\\*\\|supercite\\|autocite\\|Autocite\\|autocite\\*\\|Autocite\\*\\|Citeauthor\\*\\|citetitle\\|citetitle\\*\\|citedate\\|citedate\\*\\|citeurl\\|fullcite\\|footfullcite\\|notecite\\|Notecite\\|pnotecite\\|Pnotecite\\|fnotecite\\|cites\\|Cites\\|parencites\\|Parencites\\|footcites\\|footcitetexts\\|smartcites\\|Smartcites\\|textcites\\|Textcites\\|supercites\\|autocites\\|Autocites\\|bibentry\\):\\([a-zA-Z0-9-_:\\./]+,?\\)+"
   .......................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................................^
org-ref-core.el:720:46: In call to replace-regexp-in-string: Unescaped literal `*' (pos 0)
  "*"
   ^
org-ref-core.el:729:3: In org-ref-label-re: Literal `-' not first or last in character alternative (pos 18)
  "label:\\([a-zA-Z0-9-_:]+,?\\)+"
   ...................^
org-ref-core.el:734:3: In org-ref-ref-re: Literal `-' not first or last in character alternative (pos 23)
  "\\(eq\\)?ref:\\([a-zA-Z0-9-_:]+,?\\)+"
   ..........................^
org-ref-core.el:1725:3: In org-ref-label-regexps: Escaped non-special character `l' (pos 0)
  "\\label{\\(?1:[+a-zA-Z0-9:\\._-]*\\)}"
   ^
org-ref-core.el:2141:55: In call to replace-regexp-in-string: Literal `-' not first or last in character alternative (pos 10)
  "[a-zA-Z0-9-_]*,"
   ..........^
org-ref-core.el:2193:17: In call to re-search-forward: Repetition of expression matching an empty string (pos 23)
  "\\\\addbibresource{\\(.*\\)?}"
   ...........................^
org-ref-core.el:3106:33: In call to re-search-forward: Escaped non-special character `a' (pos 0)
  "\\attachfile{\\([^}]*\\)}"
   ^
org-ref-helm.el:303:19: In call to re-search-forward: Ranges `A-z' and `a-z' overlap (pos 17)
  "[^#+]label:\\([a-zA-z0-9:-]*\\)"
   ..................^
org-ref-helm.el:322:39: In call to re-search-forward: Ranges `A-z' and `a-z' overlap (pos 14)
  "\\\\label{\\([a-zA-z0-9:-]*\\)}"
   .................^
org-ref-isbn.el:174:32: In call to re-search-forward: Literal `-' not first or last in character alternative (pos 34)
  "@[a-zA-Z]+{.+\\(\n +[a-z ={.,?:;'0-9-\n}]+\\)+}$"
   ....................................^
org-ref-latex.el:36:50: In call to replace-regexp-in-string: Unescaped literal `*' (pos 0)
  "*"
   ^