Fix doi-utils-add-bibtex-entry-from-doi when no trailing newline
doi-utils-add-bibtex-entry-from-doi will fail to add an bibtex entry if the bib
file does not have ending newline, and report end of buffer. There is no point
in running (forward-char 1) after (goto-char (point-max)), since you are already
at the end of the document. The bug was introduced in commit b12b8f05.