partial fix for issue #746
When you have bib files that are not part of the default setup, it is difficult
to use notes. this hacks around it, making a bib file temporarily part of the
default setup. It doesn't fix the issue in the notes file, which is you have to
specify a bib file there.