Fix for issue #718
this enables cite links inside keywords. Before this did not work because I used
org-parsing, which doesn't work on a keyword which parses as a keyword. This
solution was proposed by @tshu-w. It works for getting the key under the cursor,
and should work for this use.

It doesn't work for me yet, but for independent reasons I think. I am unable to
open a cite link from a string due to a fontification error. So far I have not
been able to figure why, somewhere the cite link fontification goes past the end
of the buffer.