Working towards a fix for issue #801
This is still tricky. I added a lot of debug statements to help see what is
happening. There were a few issues for the case described here.

1. the pop was not working right, that is commented out now.
2. the change functions did not handle insertion totally right. they are better
now.

I separated out some functions for insertion and deletion events which made it
easier to see what was happening.

I don't know of any current issues with this code, so I am committing it.