probable fix for issue #845
This issue is probably a corner case when there are $ in an entry. s-format
seems to pick that up as a field. It does not seem to happen when there is an
orhc-candidate-format defined.

The solution here is to move the case where
there is no format defined outside of s-format so it cannot be triggered by it.