Use org-ref-url-html-to-bibtex for web pages when no DOI is found
Relatively recently, org-ref-url-html-to-bibtex was added to attempt
to add a best guess bibliography entry for a URL dropped into a
bibtex-mode buffer. Previously, a default "@misc" entry was added.

This change replaces the code that always inserted @misc with a call
to the newer function that provides more useful detail.

Also changed the default value for "Author" when no author is found in
the HTML to "Unknown" instead of "nil".

As a side effect of the change, @online is used for BibLaTeX files,
since that's how org-ref-html-to-bibtex works.