comment out the trailing code that deletes punctuation.
I don't recall why this is here. It seems incorrect (issue #877), but I guess
was there for some reason in the past, maybe for a case where you add pre/post
text to a citation in a sequence.

This is only used in the helm-bibtex backend as far as I can tell.