Fix OOB problem with org-ref-delete-labels
* org-ref-core.el (org-ref-delete-labels): Temporarily widen buffer before
grabbing text-properties

`get-text-property' is used at many points within modules of
`org-ref-delete-labels', but it errors when trying to grab the properties
beyond the buffer’s restriction (i.e. narrowing).  This is particularly
problematic for a function that hooks onto `before-change-functions'.

Saving the restriction and widening the before resolves the issue, but
considering the call-frequency of `before-change-functions', refactoring will
have to be considered down the line.