partial fix for issue #914
I need to have a way to pass a prefix arg to the next command during completion.
In ivy this is done via ivy-current-prefix-arg, but this library should be
completion agnostic.

I have defined the variable in the library so it should be nil at least for
non-ivy users, but that means they can't use the prefix arg to change the ref
type either. I am not sure what the solution is for them. ivy does this by setq
the ivy-current-prefix-arg when you choose an action.

This also fixes issue #912 I think, where the label context was missing.