capture C-u during completion (issue #914)
This adds a capture mechanism similar to what ivy does.

I could not get any variation of current or last-prefix-arg to work during
completion. This method uses a variable to capture that, and a minibuffer hook
with a local keybinding to capture prefix args for use in subsequent code. seems
like a lot of work to avoid a dependency.