get rid of cache
I think this isn't working reliably on MELPA. Something isn't right with lexical
scoping perhaps when it is compiled. It worked fine without compiling.

I can't tell a difference in performance without it right now, so I am taking it
out to be simple. there maybe a way to use the cache from bibtex-completion, but
for now this seems ok