attempt to add elements from issue #921
It seems this is trickier than anticipated. I added these as special blocks,
which makes more sense than as latex, since they are more org like. but, if you
write them as latex environments, they will not get found. I think I might have
to break up the referenceables into a list of functions so it is easier to use.
Each type will have a function that returns a label if it is that type and
referenceable.