fix for issue #971
this expands include files in the copy of the buffer. That is necessary to get
citations in included files handled correctly.

Hopefully this doesn't mess up subtree exports too though. I am using a marker
to save the place where the cursor was, and hopefully this still does the right thing.