add a cache for bst styles to speed up after initial check
In issue #967 @novusshu reports slow editing on the bibliography and
bibliographystyle lines. In the case of the style line it is probably because it
would run kpsewhich on every edit, which might be slow especially on windows.
The result of that isn't likely to change, so it makes sense to cache that
result instead, and just use that.