Probable fix for issue #966
The problem was that if you have a doi in your bibtex file like
10.1007/3-540-32144-6_8 then try to add one with a doi like
10.1007/3-540-32144-6, then the function would match the sub-pattern in the
first one, and error out. I added a bit to the end of the doi so that it would
only match if it is at the end of the symbol. you cannot use the end of a word,
because _ is not a word character.