address some issues uncovered in issue #1041
There is a corner case where a ; is in the prefix note. ; is used as a
separator, so I don't know a good way to handle this without escaping it
somehow. I don't do that here.

Instead, I join the global and local prefix in this case, which achieves the
same goal.

It also appears you may need an empty suffix placeholder with just a prefix, and
I think this commit fixes that too.