Added support to copy/move PDF from citation reference.
This patch improves the functionality when typing return on a reference.

It adds the option to copy/add the PDF to the current citation.  If the
citation already has a PDF, a message is issued and nothing happens.

This commit adds two functions:

  org-ref-library-path
     "return the library path".
     This function has been refactored from org-ref-pdf-to-bibtex

   org-ref-add-pdf-at-point
      "Add the pdf for bibtex key under point if it exists."

And add the new functionality to org-ref-citation-hydra and modifies
org-ref-pdf-to-bibtex to use the refactored org-ref-library-path.