fix error <path/a.bib> <path/b.bib> are the same file on open URL
(org-ref-find-bibliography) can return path from config variable, which may contain symbolic links in the path, when it is compared with (org-ref-get-bibtex-key-and-file), which uses (buffer-file-name) to retrieve a bib file path, it returns an absolute path with all symbolic links resolved. Two different paths pointing to the same file leads to the above error and prevents the body of the let clause from executing. Using the file-truename prevents this error.