Attempt to improve performance (see issue #1120)
I have tried to make it only parse the document once, and then use a cache on
subsequent lookups. I found an error in the cache where it wasn't using the
right test for equal that is fixed now. I also store data on text properties
now, so it should be much faster to compute faces and tooltips.

Some things are not quite as good as I want, like following links is not as
precise as I would like, and may still need some fine tuning.

hopefully this is more performant in large files.